Tic_Tac_Toe
The TIC TAC TOE game is a game for two players “X” & “O”, who takes turns making the spaces in a 3x3 grid. The player who succeeds in placing three respective marks in a horizontal, vertical or diagonal row wins the game.

The game is offered in two different modes:
1.	 Player VS Player
2.	 Player VS Computer